pyproject.toml
[project]
name = "pr-auto-reviewer"
version = "0.1.0"
description = "Automatically fix code based on PR review comments"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyGithub>=2.1.1",
    "anthropic>=0.40.0",
    "google-cloud-aiplatform>=1.38.1",
    "python-dotenv>=1.0.0",
    "GitPython>=3.1.40",
    "click>=8.1.7",
    "rich>=13.7.0",
    "questionary>=2.0.1",
]

[project.scripts]
pr-reviewer = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "config", "code_modifier", "code_reviewer", "git_ops", "github_client"]

[tool.setuptools.packages.find]
include = ["models*"]

[tool.uv]
dev-dependencies = []
